person-of-interest/mcp_server/requirements.txt
# ── MCP Framework ────────────────────────────────────────────────────────────
mcp>=1.0.0

# ── GitHub ────────────────────────────────────────────────────────────────────
PyGithub>=2.1.1

# ── Jira ──────────────────────────────────────────────────────────────────────
atlassian-python-api>=3.41.0

# ── Redis ─────────────────────────────────────────────────────────────────────
redis>=5.0.0

# ── MQTT ──────────────────────────────────────────────────────────────────────
paho-mqtt>=2.1.0

# ── HTTP client (FAISS/POI backend calls) ─────────────────────────────────────
requests>=2.31.0

# ── OpenVINO + CV (optional — skip if models not available) ──────────────────
openvino>=2024.0
opencv-python-headless>=4.10.0
numpy>=1.26.0

# ── Docker SDK ────────────────────────────────────────────────────────────────
docker>=7.0.0
